+++ py-polars/polars/_reexport.py
@@ -1,7 +1,7 @@
 """Re-export Polars functionality to avoid cyclical imports."""
 
 from polars.dataframe import DataFrame
-from polars.expr import Expr
+from polars.expr import Expr, When
 from polars.lazyframe import LazyFrame
 from polars.series import Series
 
@@ -10,4 +10,5 @@ __all__ = [
     "Expr",
     "LazyFrame",
     "Series",
+    "When",
 ]
