+++ py-polars/polars/expr/__init__.py
@@ -1,5 +1,7 @@
 from polars.expr.expr import Expr
+from polars.expr.whenthen import When
 
 __all__ = [
     "Expr",
+    "When",
 ]
