+++ py-polars/polars/functions/whenthen.py
@@ -1,21 +1,19 @@
 from __future__ import annotations
 
 import contextlib
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
 import polars._reexport as pl
 from polars.utils._parse_expr_input import parse_as_expression
-from polars.utils._wrap import wrap_expr
 
 with contextlib.suppress(ImportError):  # Module not available when building docs
     import polars.polars as plr
 
 if TYPE_CHECKING:
-    from polars import Expr
     from polars.type_aliases import IntoExpr
 
 
-def when(expr: IntoExpr) -> When:
+def when(expr: IntoExpr) -> pl.When:
     """
     Start a "when, then, otherwise" expression.
 
@@ -95,86 +93,4 @@ def when(expr: IntoExpr) -> When:
     """
     expr = parse_as_expression(expr)
     pywhen = plr.when(expr)
-    return When(pywhen)
-
-
-class When:
-    """Utility class. See the `when` function."""
-
-    def __init__(self, pywhen: Any):
-        self._pywhen = pywhen
-
-    def then(self, expr: IntoExpr) -> WhenThen:
-        """
-        Values to return in case of the predicate being `True`.
-
-        See Also
-        --------
-        pl.when : Documentation for `when, then, otherwise`
-
-        """
-        expr = parse_as_expression(expr, str_as_lit=True)
-        pywhenthen = self._pywhen.then(expr)
-        return pl.WhenThen(pywhenthen)
-
-
-class WhenThen(pl.Expr):
-    """Utility class. See the `when` function."""
-
-    def __init__(self, pywhenthen: Any):
-        self._pywhenthen = pywhenthen
-        self._pyexpr = pywhenthen.otherwise(None)
-
-    def when(self, predicate: IntoExpr) -> WhenThenThen:
-        """Start another "when, then, otherwise" layer."""
-        predicate = parse_as_expression(predicate)
-        return WhenThenThen(self._pywhenthen.when(predicate))
-
-    def otherwise(self, expr: IntoExpr) -> Expr:
-        """
-        Values to return in case of the predicate being `False`.
-
-        See Also
-        --------
-        pl.when : Documentation for `when, then, otherwise`
-
-        """
-        expr = parse_as_expression(expr, str_as_lit=True)
-        return wrap_expr(self._pywhenthen.otherwise(expr))
-
-
-class WhenThenThen(pl.Expr):
-    """Utility class. See the `when` function."""
-
-    def __init__(self, pywhenthenthen: Any):
-        self.pywhenthenthen = pywhenthenthen
-        self._pyexpr = pywhenthenthen.otherwise(None)
-
-    def when(self, predicate: IntoExpr) -> WhenThenThen:
-        """Start another "when, then, otherwise" layer."""
-        predicate = parse_as_expression(predicate)
-        return WhenThenThen(self.pywhenthenthen.when(predicate))
-
-    def then(self, expr: IntoExpr) -> WhenThenThen:
-        """
-        Values to return in case of the predicate being `True`.
-
-        See Also
-        --------
-        pl.when : Documentation for `when, then, otherwise`
-
-        """
-        expr = parse_as_expression(expr, str_as_lit=True)
-        return WhenThenThen(self.pywhenthenthen.then(expr))
-
-    def otherwise(self, expr: IntoExpr) -> Expr:
-        """
-        Values to return in case of the predicate being `False`.
-
-        See Also
-        --------
-        pl.when : Documentation for `when, then, otherwise`
-
-        """
-        expr = parse_as_expression(expr, str_as_lit=True)
-        return wrap_expr(self.pywhenthenthen.otherwise(expr))
+    return pl.When(pywhen)
